Обновляет reducer, ActionType, ActionCreator
