+++ src/redux/action-type.js
@@ -1,0 +1,4 @@
+export const ActionType = {
+  CHANGE_GENRE: `genreFilterChange`,
+  FILTER_FILMS_BY_GENRE: `filterFilmsByGenre`
+};
